pptx2md  #69
As mentioned in Issue #69, here is my PR :)

python-pptx will throw a ValueError, if the lazily loaded file extension of an image is not a supported type (e.g. MPO)
    - ref. parser.py:155
This ValueError does not get caught while processing the shapes which in turn will crash the program . The simple solution is to just add ValueError to the exception tuple where it gets called in the process_shapes function.
The other place the function gets called already has a try-except block and simply ignores all exceptions